minor refactor of things
